This PR fixes the bug when exporting OBB model for TFLite with INT8 quantization. During normalization dist2bbox is used instead of dist2rbox.

NOTE: tests have been executed using the validation set. The test set is missing the labels. TFLite quantization was set to per-channel.

|Model|mAP50|mAP50:95|Note|
|--|--|--|--|
|Before yolov8n-obb.pt| 0.468 | 0.283||
|Before yolov8n-obb.TFLite INT8| 0.375 | 0.245||
|Before yolov8n-obb.TFLite Full INT| 0.305|0.175||
|After yolov8n-obb.pt| 0.468 | 0.283| |
|After yolov8n-obb.TFLite INT8| 0.496|0.355| How? |
|After yolov8n-obb.TFLite Full INT| 0.394|0.244||

Reproduce tests:
```
yolo export model=yolov8n-obb.pt data=coco128.yaml imgsz=640 format=tflite int8
yolo val obb model=yolov8n-obb.pt data=DOTAv1.yaml device=0 imgsz=640 batch=4 split=val
yolo val obb model=yolov8n-obb_saved_model/yolov8n-obb_int8.tflite data=DOTAv1.yaml device=0 imgsz=640 int8 split=val
yolo val obb model=yolov8n-obb_saved_model/yolov8n-obb_full_integer_quant.tflite data=DOTAv1.yaml device=0 imgsz=640 int8 split=val
```

It is surprising to observe that the TFLite INT8, where only weights are INT8 and inference is in FP32, is performing better than the original PyTorch model. This should be further tested, but I am not familiar with the OBB neither with DOTA dataset. Theoretically the only difference is that the bbox coordinates are normalized. Could it have an influence on the bbox rotations?


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Refactor of bounding box decoding in neural network heads to improve flexibility.

### 📊 Key Changes
- `decode_bboxes` method now takes `anchors` as an argument instead of using `self.anchors`.
- Bounding box decoding operation has been updated within the `forward` method of the network to better handle different conditions, such as when exporting models to TensorFlow Lite or Edge TPU.
- Calls to `dist2bbox` and `dist2rbox` now directly pass calculated bounding boxes and anchors.

### 🎯 Purpose & Impact
- 💡 **Purpose**: The changes aim to improve the stability and flexibility of bounding box decoding across different network heads and export formats.
- 🔍 **Impact**: Developers will have an easier time adapting and maintaining the bounding box decoding logic for different use cases, potentially resulting in better model compatibility and performance when exporting to various platforms.
- 🚀 **To users**: These updates might lead to more accurate object detections and support for broader deployment scenarios, especially for edge computing devices.