@glenn-jocher 2024-02-03
@adamp87 based on the table it seems like something may be wrong in the PR, as it should not be possible to increase mAP beyond the base PyTorch model.

@Laughing-q what do you think?