@Laughing-q 2024-02-04
@adamp87 Thanks for the fix!
@glenn-jocher @adamp87 The map from torch model is lower because we default use rectangle inference when doing validation with torch model, but for exported model it's square inference. The map looks fine when I manually set `rect=False`.
```bash
yolo val obb model=yolov8n-obb.pt data=DOTAv1.yaml device=0 imgsz=640 batch=4 split=val
```
```bash
                   all        458      28853      0.698       0.43      0.468      0.283
                 plane        458       2531      0.804      0.629      0.668      0.414
                  ship        458       8960      0.735      0.478      0.555      0.286
          storage tank        458       2888      0.847      0.231      0.347      0.182
      baseball diamond        458        214      0.776      0.435      0.525       0.34
          tennis court        458        760       0.91      0.883      0.912      0.704
      basketball court        458        132      0.637      0.348      0.411       0.29
    ground track field        458        144      0.585      0.361      0.391      0.272
                harbor        458       2090      0.565      0.537      0.503       0.22
                bridge        458        464      0.606      0.127      0.149     0.0507
         large vehicle        458       4387       0.76      0.767      0.784      0.471
         small vehicle        458       5438       0.66      0.496      0.545       0.33
            helicopter        458         73      0.767      0.288      0.334      0.181
            roundabout        458        179      0.617      0.128       0.16     0.0962
     soccer ball field        458        153      0.553      0.392       0.37      0.229
         swimming pool        458        440      0.646      0.344      0.363      0.175
```
```bash
yolo val obb model=yolov8n-obb.pt data=DOTAv1.yaml device=0 imgsz=640 batch=4 split=val rect=False
```
```bash
                   all        458      28853      0.729      0.447      0.496      0.356
                 plane        458       2531      0.853      0.672      0.724      0.537
                  ship        458       8960      0.766      0.494      0.574      0.351
          storage tank        458       2888      0.863      0.234      0.365      0.228
      baseball diamond        458        214      0.764       0.43      0.525      0.393
          tennis court        458        760      0.908      0.882      0.917      0.846
      basketball court        458        132      0.642      0.356       0.42      0.349
    ground track field        458        144      0.583      0.389      0.388      0.327
                harbor        458       2090       0.74      0.704       0.72      0.398
                bridge        458        464      0.658      0.137      0.166      0.058
         large vehicle        458       4387      0.793       0.79      0.819      0.608
         small vehicle        458       5438      0.676       0.49      0.551      0.358
            helicopter        458         73      0.887      0.329      0.404       0.28
            roundabout        458        179      0.644      0.123      0.156      0.112
     soccer ball field        458        153      0.507      0.346      0.365       0.31
         swimming pool        458        440      0.655      0.336      0.353      0.185
```