@adamp87 2024-02-04
@Laughing-q thanks for the clarification. It was bugging me also!
@glenn-jocher I pushed one more commit. I changed only variables names during normalization, where the normalization is done with the width and height of the grid not the input image size.