@adamp87 2024-02-04
@glenn-jocher one more note, OpenVINO INT8 quantization fails on main branch with `nncf: list index out of range`. I just pulled the repo now. The suggested solution in https://github.com/ultralytics/ultralytics/pull/7516#issuecomment-1924412632 fixes this problem. 

`yolo export model=yolov8n-obb.pt data=coco128.yaml imgsz=640 format=openvino int8`

|Model|mAP50|mAP50:95|Note|
|--|--|--|--|
|After yolov8n-obb.pt| 0.468 | 0.283| |
|After yolov8n-obb.TFLite INT8| 0.489 | 0.349 |  |
|After yolov8n-obb.TFLite Full INT| 0.394 | 0.244|  |
|OpenVINO INT8 | N/A | N/A | Main branch fails |
|OpenVINO INT8 | 0.493| 0.352 | https://github.com/ultralytics/ultralytics/pull/7516#issuecomment-1924412632 |