@glenn-jocher 2024-02-04
@adamp87 Thanks for bringing this to our attention. We're aware of the issue with OpenVINO INT8 quantization on the main branch and appreciate the link to the suggested solution. We'll review the proposed changes and work on integrating a fix. Your efforts to improve the repo are invaluable. 🛠️ Keep an eye on the updates!