@glenn-jocher 2024-02-06
@adamp87 thank you for your efforts here! I've merged this into main now. I was reviewing https://github.com/ultralytics/ultralytics/pull/7516#issuecomment-1924412632 but we can't merge this as it's failing OpenVINO Benchmarks CI on main. 

Now that I've merged this to main I will re-run the CI there to see if this now passes.