@adamp87 2024-02-06
@glenn-jocher thanks for the update. 
Im confused on the https://github.com/ultralytics/ultralytics/pull/7516#issuecomment-1924412632 comment. Do you mean https://github.com/ultralytics/ultralytics/pull/7516#issuecomment-1924412632 or the PR https://github.com/ultralytics/ultralytics/pull/7516? They are two different things. I havent merged https://github.com/ultralytics/ultralytics/pull/7516#issuecomment-1924412632 into the PR its just in my own fork dangling.